feat: new modules and authentication added

+++ src/modules/projects/projects.service.ts
@@ -2,18 +2,29 @@ import { Injectable } from '@nestjs/common'
 import { ProjectRequestDTO } from './projects.dto'
 import { PrismaService } from 'src/prisma.service'
 import { CollaboratorRole } from '@prisma/client'
+import { RequestContextService } from 'src/common/services/request-context/request-context.service'
 
 @Injectable()
 export class ProjectsService {
-  constructor(private readonly prisma: PrismaService) {}
+
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly requestContext: RequestContextService
+  ) {}
 
   findAll() {
-    return this.prisma.project.findMany()
+    const userId = this.requestContext.getUserId()
+    return this.prisma.project.findMany({
+      where: {
+        createdById: userId,
+      }
+    })
   }
 
   findById(id: string) {
+    const userId = this.requestContext.getUserId()
     return this.prisma.project.findFirst({
-      where: { id },
+      where: { id, createdById: userId },
       select: {
         id: true,
         name: true,
@@ -37,17 +48,18 @@ export class ProjectsService {
   }
 
   async create(data: ProjectRequestDTO) {
+    const userId = this.requestContext.getUser()?.id 
     const project = await this.prisma.project.create({
       data: {
         ...data,
-        createdById: '123',
+        createdById: userId,
       },
     })
 
     await this.prisma.projectCollaborator.create({
       data: {
         projectId: project.id,
-        userId: '123',
+        userId: userId,
         role: CollaboratorRole.OWNER,
       },
     })
@@ -56,11 +68,13 @@ export class ProjectsService {
   }
 
   update(id: string, data: ProjectRequestDTO) {
-    return this.prisma.project.update({ where: { id }, data })
+    const userId = this.requestContext.getUserId()
+    return this.prisma.project.update({ where: { id, createdById: userId }, data })
   }
 
   async remove(id: string) {
+    const userId = this.requestContext.getUserId()
     await this.prisma.task.deleteMany({ where: { projectId: id } })
-    return this.prisma.project.delete({ where: { id } })
+    return this.prisma.project.delete({ where: { id, createdById: userId } })
   }
 }
